+++ R/plotlib.R
@@ -80,6 +80,7 @@ plot_confint <- function (p, data = NULL, lower = 'lower', upper = 'upper',
 #' @param label.hjust Horizontal adjustment for text labels
 #' @param label.vjust Vertical adjustment for text labels
 #' @param label.repel Logical flag indicating whether to use \code{ggrepel}, enabling this may take some time for plotting
+#' @param label.show.legend Logical value indicating whether to show the legend of the text labels
 #' @return ggplot
 plot_label <- function(p, data, x = NULL, y = NULL, label = TRUE, label.label = 'rownames',
                        label.colour = NULL, label.alpha = NULL,
@@ -87,7 +88,7 @@ plot_label <- function(p, data, x = NULL, y = NULL, label = TRUE, label.label = 
                        label.family = NULL, label.fontface = NULL,
                        label.lineheight = NULL,
                        label.hjust = NULL, label.vjust = NULL,
-                       label.repel = FALSE) {
+                       label.repel = FALSE, label.show.legend = NA) {
 
   if (!is.data.frame(data)) {
     stop(paste0('Unsupported class: ', class(data)))
@@ -115,7 +116,8 @@ plot_label <- function(p, data, x = NULL, y = NULL, label = TRUE, label.label = 
                           size = label.size, angle = label.angle,
                           family = label.family, fontface = label.fontface,
                           lineheight = label.lineheight,
-                          hjust = label.hjust, vjust = label.vjust)
+                          hjust = label.hjust, vjust = label.vjust,
+                          show.legend = label.show.legend)
   }
   p
 }
@@ -489,6 +491,7 @@ autoplot.ggmultiplot <- function(object, ...) {
 #' @param loadings.label.hjust Horizontal adjustment for loadings text labels
 #' @param loadings.label.vjust Vertical adjustment for loadings text labels
 #' @param loadings.label.repel Logical flag indicating whether to use \code{ggrepel} automatically
+#' @param label.show.legend Logical value indicating whether to show the legend of text labels
 #' @param frame Logical value whether to draw outliner convex / ellipse
 #' @param frame.type Character specifying frame type.
 #' 'convex' or types supporeted by \code{ggplot2::stat_ellipse} can be used.
@@ -528,6 +531,7 @@ ggbiplot <- function(plot.data, loadings.data = NULL,
                      loadings.label.hjust = NULL,
                      loadings.label.vjust = NULL,
                      loadings.label.repel = FALSE,
+                     label.show.legend = NA,
                      frame = FALSE, frame.type = NULL,
                      frame.colour = colour, frame.level = 0.95,
                      frame.alpha = 0.2,
@@ -560,7 +564,8 @@ ggbiplot <- function(plot.data, loadings.data = NULL,
                   label.lineheight = label.lineheight,
                   label.hjust = label.hjust,
                   label.vjust = label.vjust,
-                  label.repel = label.repel)
+                  label.repel = label.repel,
+                  label.show.legend = label.show.legend)
 
   if (loadings.label && !loadings) {
     # If loadings.label is TRUE, draw loadings
@@ -592,7 +597,8 @@ ggbiplot <- function(plot.data, loadings.data = NULL,
                     label.lineheight = loadings.label.lineheight,
                     label.hjust = loadings.label.hjust,
                     label.vjust = loadings.label.vjust,
-                    label.repel = loadings.label.repel)
+                    label.repel = loadings.label.repel,
+                    label.show.legend = label.show.legend)
   }
 
   if (missing(frame) && !is.null(frame.type)) {